docker_image/requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.8.0
asttokens==3.0.0
blinker==1.9.0
blis==1.2.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
confection==0.1.5
contourpy==1.3.1
contractions==0.1.73
cycler==0.12.1
cymem==2.0.11
databricks-sdk==0.44.1
debugpy==1.8.12
decorator==5.2.0
Deprecated==1.2.18
dnspython==2.7.0
docker==7.1.0
email_validator==2.2.0
executing==2.2.0
fastapi==0.115.8
fastapi-cli==0.0.7
fastrlock==0.8.3
filelock==3.17.0
Flask==3.1.0
fonttools==4.56.0
fsspec==2025.2.0
gensim==4.3.3
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-core==2.4.2
google-cloud-storage==3.0.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.68.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.29.1
idna==3.10
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.32.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
Mako==1.3.9
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mlflow==2.20.2
mlflow-skinny==2.20.2
mpmath==1.3.0
murmurhash==1.0.12
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
numpy==1.26.4
opentelemetry-api==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.1.0
pip==25.0
platformdirs==4.3.6
preshed==3.0.9
prompt_toolkit==3.0.50
proto-plus==1.26.0
protobuf==5.29.3
psutil==7.0.0
pure_eval==0.2.3
pyahocorasick==2.1.0
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.13.1
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.38
sqlparse==0.5.3
srsly==2.5.1
stack-data==0.6.3
starlette==0.45.3
sympy==1.13.1
textsearch==0.0.24
thinc==8.3.4
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
waitress==3.0.2
wasabi==1.1.3
watchfiles==1.0.4
wcwidth==0.2.13
weasel==0.4.1
websockets==15.0
Werkzeug==3.1.3
wheel==0.45.1
wordcloud==1.9.4
wrapt==1.17.2
zipp==3.21.0
